+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - added function `var_dump()`
 - added `ns` alias for `namespace` keyword (GH-51)
 - added `format_args()` function
+- added `webserver` module
 
 #### Bug Fixes
 - fixed a bug in `eval` function